729: Adding FetchTrustedDirectoryFilter (#264)
* 729: Adding FetchTrustedDirectoryFilter

This filter will add the TrustedDirectory for this ApiClient to the attributes context. NOTE: FetchTrustedDirectoryFilter should be installed in a filter chain after the FetchApiClientFilter filter, as the ApiClient data is required to determine the TrustedDirectory.

Added filter chain: FetchApiClientAndTrustedDirectoryDataChain which can be used to fetch the ApiClient and then the TrustedDirectory. Updating existing usages of the FetchApiClientFilter in routes to use the new chain.

https://github.com/SecureApiGateway/SecureApiGateway/issues/729

* 729: Extracted IdmApiClientDecoder from FetchApiClientFilter

Extracted the logic to decode the IDM json response into an ApiClient. Improved decode logic to handle errors more gracefully.

Fixed bug which treated jwksUri as a mandatory field, when it is optional. Added helper methods which create IDM test data with all fields set and another method which only sets the mandatory fields.

https://github.com/SecureApiGateway/SecureApiGateway/issues/729