831: Validate MTLS client cert when accessing token_endpoint (#316)
* 831: Extracting IdmApiClientService from FetchApiClientFilter

Extracting the logic to fetch the ApiClient into the IdmApiClientService. This service can then be used in other filters.

https://github.com/SecureApiGateway/SecureApiGateway/issues/842

* 831: Extracting ApiClientJwkSetService which implements the logic to get the JWKS for an ApiClient

Updating FetchApiClientJwksFilter to use ApiClientJwkSetService

https://github.com/SecureApiGateway/SecureApiGateway/issues/842

* 831: Adding helper method to TrustedDirectoryService which get the trusted directory based on an ApiClient

https://github.com/SecureApiGateway/SecureApiGateway/issues/831

* 831: Extracting ApiClientService interface from IdmApiClientService

https://github.com/SecureApiGateway/SecureApiGateway/issues/831

* 831: Adding CertificateRetriever and FromHeaderCertificateRetriever

FromHeaderCertificateRetriever resolves the client's mTLS certificate from a HTTP request header. The resolver will decode the certificate and return an X509Certificate object, or it will throw an exception.

Deprecating the CertificateFromHeaderSupplier which does a similar job but only returns the raw header string, leaving it up to the caller to parse the cert.

The new FromHeaderCertificateRetriever can be used to reduce code duplication involved in parsing the cert in multiple places.

https://github.com/SecureApiGateway/SecureApiGateway/issues/831

* 831: Initial implementation of TokenEndpointTransportCertValidationFilter

TokenEndpointTransportCertValidationFilter is a specialised form of TransportCertValidationFilter, this filter works on the response path and fetches the data needed to do the validation by calling services directly (rather than accessing values from the context). This is required because for the token endpoint route, we do not know that the client has been authenticated until we get the response back from AM, therefore it is safest do the validation on the response path.

https://github.com/SecureApiGateway/SecureApiGateway/issues/831