879: Using IG_TEST_DIRECTORY_ENABLED config in production (#341)
* 879: Using IG_TEST_DIRECTORY_ENABLED config in production

Allow IG_TEST_DIRECTORY_ENABLED env var to configure this feature in production images. This allows us to run a production image in our development environments with this feature enabled.

By default, this feature is turned off.

Currently, security.enableTestTrustedDirectory is only used as part of the DCR route. Expanding usage to the jwkms routes.

https://github.com/SecureApiGateway/SecureApiGateway/issues/879