859: Check route role against roles in ApiClient (#342)
* 859: Check route role against roles in ApiClient

The CertificateCheck.groovy script used the roles from the presented
eidas certificate. This restricts the API client to those that have an
eidas style certificate. We now use the roles specified in the software
statement at the time the client performed dynamic client registration
to determine the roles that the client is allowed to perform.

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/859

---------

Co-authored-by: Dan Badham <dan.badham@forgerock.com>