1121: Making use of TokenRequestHandler only in IDMClientHandler (#427)
* 1121: Removing request for gateway access_token from OB Account Access Consent route

A gateway client_credentials access_token was being requested as part of the OB Account Access Consent route, the token is not needed.

The route is reverse proxying the RS, and checks the ApiClient is providing a client_credentials access_token. No additional gateway token is required.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1121

* 1121: Moving TokenRequestHandler into IDMClientHandler

TokenRequestHandler is only needed by the IDMClientHandler. Moving it as an inline handler definition so that it cannot be used elsewhere.

Changing the gateway token request from the token endpoint route to use the ForgeRockClientHandler, as it can use client_credentials flow.

Renaming SettingNewEntity => SetIdmPasswordGrantCredentials to make it clearer what the script does.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1121